% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelops_handlers.R
\name{handler_startup.train}
\alias{handler_startup.train}
\alias{handler_predict.train}
\title{caret model handler functions for API endpoint}
\usage{
\method{handler_startup}{train}(modelops, ...)

\method{handler_predict}{train}(modelops, ...)
}
\arguments{
\item{modelops}{A deployable model object created with \code{\link[modelops:modelops]{modelops()}}}

\item{...}{Other arguments passed to \code{predict()}, such as prediction \code{type}}
}
\description{
caret model handler functions for API endpoint
}
